Relax the requirement to use a list when defining a constant
constants required the value to be a list if it was an expression, i.e.
constants [[a [1 + 2 + 3]]]

Now, the expression can just follow the name, i.e. constants [[a 1 + 2 + 3]]
For the compatibility with javalogo version, lists are accepted as well.

Also, fixes the exception, caused by the lack of state.name during pass1.