Improve the response time of an interactive command by not waiting too much
The main issue was that mbedlogo.read actually waits at least timeout
before return. In attempt to return as much data as possible, it waits an
extra timeout to be sure that no more data is coming in. It slows down the
response to an interactive command.

Now, if we know that we're expecting only one byte in return, we just call
the mbedlogo.reader.read directly.

On top of that, timeouts are reduced, but they still appear working.